move RegisterGenerator so generators still get registered
