+++ corehq/apps/repeaters/models.py
@@ -134,57 +134,6 @@ class GeneratorCollection(object):
         return self.format_generator_map[format].generator_class
 
 
-class RegisterGenerator(object):
-    """Decorator to register new formats and Payload generators for Repeaters
-
-    args:
-        repeater_cls: A child class of Repeater for which the new format is being added
-        format_name: unique identifier for the format
-        format_label: description for the format
-
-    kwargs:
-        is_default: whether the format is default to the repeater_cls
-    """
-
-    generators = {}
-
-    def __init__(self, repeater_cls, format_name, format_label, is_default=False):
-        self.format_name = format_name
-        self.format_label = format_label
-        self.repeater_cls = repeater_cls
-        self.label = format_label
-        self.is_default = is_default
-
-    def __call__(self, generator_class):
-        if not self.repeater_cls in RegisterGenerator.generators:
-            RegisterGenerator.generators[self.repeater_cls] = GeneratorCollection(self.repeater_cls)
-        RegisterGenerator.generators[self.repeater_cls].add_new_format(
-            self.format_name,
-            self.format_label,
-            generator_class,
-            is_default=self.is_default
-        )
-        return generator_class
-
-    @classmethod
-    def generator_class_by_repeater_format(cls, repeater_class, format_name):
-        """Return generator class given a Repeater class and format_name"""
-        generator_collection = cls.generators[repeater_class]
-        return generator_collection.get_generator_by_format(format_name)
-
-    @classmethod
-    def all_formats_by_repeater(cls, repeater_class, for_domain=None):
-        """Return all formats for a given Repeater class"""
-        generator_collection = cls.generators[repeater_class]
-        return generator_collection.get_all_formats(for_domain=for_domain)
-
-    @classmethod
-    def default_format_by_repeater(cls, repeater_class):
-        """Return default format_name for a Repeater class"""
-        generator_collection = cls.generators[repeater_class]
-        return generator_collection.get_default_format()
-
-
 class Repeater(QuickCachedDocumentMixin, Document, UnicodeMixIn):
     """
     Represents the configuration of a repeater. Will specify the URL to forward to and
@@ -221,9 +170,11 @@ class Repeater(QuickCachedDocumentMixin, Document, UnicodeMixIn):
         return get_success_repeat_record_count(self.domain, self._id)
 
     def format_or_default_format(self):
+        from corehq.apps.repeaters.repeater_generators import RegisterGenerator
         return self.format or RegisterGenerator.default_format_by_repeater(self.__class__)
 
     def get_payload_generator(self, payload_format):
+        from corehq.apps.repeaters.repeater_generators import RegisterGenerator
         gen = RegisterGenerator.generator_class_by_repeater_format(self.__class__, payload_format)
         return gen(self)
